Topology module clean up
